feat(mc): G-1113.D.7b — iteration-import GitHub parsers behind the adapter boundary
Behaviour-preserving code-move: the GitHub `issues` webhook / `gh api`
payload parsers move out of api/iteration-import.ts into the
adapters/github boundary, so GitHub's payload schema lives with the
other GitHub-specific logic and iteration-import deals only in the
neutral metadata shapes.

- adapters/github/iteration-webhook.ts (new): IssuesWebhookEnvelope +
  ParentIssueMetadata + SubIssueMetadata types, and the parentMetadataFromWebhook
  / subIssueRefsFromWebhook / parentMetadataFromGhResponse parsers — moved
  verbatim (no parsing logic changed). Re-exported via adapters/github/index.
- api/iteration-import.ts: parsers + envelope type removed; the neutral
  metadata types re-exported from the adapter for back-compat; import
  orchestration (importIterationFromMetadata/importSubIssueFromMetadata)
  unchanged.
- api/handlers.ts + iteration-import.test.ts: import the 3 parsers from
  adapters/github directly.

No webhook HMAC / auth touched (that validation lives in the gh-webhook
tap, upstream of this parsing). Verified by the existing parser test
suite (37 pass) — behaviour preserved by construction.

Closes #599. Part of #590 / #577 / #354.
