+++ src/surface/mc/adapters/github/index.ts
@@ -20,6 +20,7 @@ import { type GitHubRef, canonicalRef, canonicalUrl } from "./ref";
 export * from "./ref";
 export * from "./fetch";
 export * from "./ingest";
+export * from "./iteration-webhook";
 
 /**
  * Normalize a parsed {@link GitHubRef} into a {@link SourceRef}. `externalId`
